requirements.txt
httpx>=0.27
pydantic>=2.0
PyYAML>=6.0
pytest>=8.0
google-adk==1.28.0
litellm==1.82.6
